#3398 update tests and move more rc-specific ones out of lfric_transformations_test
