+++ src/psyclone/tests/domain/lfric/transformations/lfric_transformations_test.py
@@ -4302,294 +4302,6 @@ def test_red_comp_w_to_n_r_clean_gt_cleaned(tmpdir):
     assert LFRicBuild(tmpdir).code_compiles(psy)
 
 
-def test_rc_no_directive():
-    ''' When the redundant computation transformation is given a Loop whose
-    parent is a directive an exception is raised as this is not
-    supported (redundant computation transformations must be applied
-    before directives are added). This test checks that this exception
-    is raised correctly.
-
-    '''
-    _, invoke = get_invoke("1_single_invoke.f90",
-                           TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Create a colouring transformation and apply this to the loop
-    ctrans = LFRicColourTrans()
-    ctrans.apply(schedule[4])
-
-    # Create an openmp transformation and apply this to the loop
-    otrans = LFRicOMPParallelLoopTrans()
-    otrans.apply(schedule[4].loop_body[0])
-
-    # Create a redundant computation transformation and apply this to the loop
-    rc_trans = LFRicRedundantComputationTrans()
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(
-            schedule[4].loop_body[0].dir_body[0], {"depth": 1})
-    assert ("Redundant computation must be applied before directives are added"
-            in str(excinfo.value))
-
-
-def test_rc_wrong_parent(monkeypatch):
-    ''' When the redundant computation transformation is given a Loop which
-    has the wrong parent, and that parent is not a Directive (which is
-    handled in a separate case) an exception is raised. This test
-    checks that this exception is raised correctly.
-
-    '''
-    _, invoke = get_invoke("1_single_invoke.f90",
-                           TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Make the parent of the loop a halo exchange
-    monkeypatch.setattr(schedule.children[4], "_parent", schedule.children[0])
-
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the loop
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(schedule.children[4], {"depth": 1})
-    assert ("the parent of the supplied loop must be the LFRicInvokeSchedule, "
-            "or a Loop") in str(excinfo.value)
-
-
-def test_rc_parent_loop_colour(monkeypatch):
-    ''' If the parent of the loop supplied to the redundant computation
-    transformation is a loop then
-
-    1) the parent loop's parent should be a schedule. If this is not
-    the case then an exception is raised.
-
-    2) the parent loop should iterate over 'colours'. If this is not
-    the case then an exception is raised.
-
-    3) the supplied loop should iterate over cells of a given
-    colour. If this is not the case then an exception is raised.
-
-    This test checks that the appropriate exceptions are correctly
-    raised for these three situations.
-
-    '''
-    _, invoke = get_invoke("1_single_invoke.f90",
-                           TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Apply colouring
-    # Create colour transformation
-    ctrans = LFRicColourTrans()
-    # Colour the loop
-    ctrans.apply(schedule.children[4])
-
-    # Make the parent of the outermost loop something other than
-    # InvokeSchedule (we use halo exchange in this case)
-    monkeypatch.setattr(schedule.children[4], "_parent", schedule.children[0])
-
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the loop
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(schedule.children[4].loop_body[0], {"depth": 1})
-    assert ("if the parent of the supplied Loop is also a Loop then the "
-            "parent's parent must be the LFRicInvokeSchedule"
-            in str(excinfo.value))
-
-    # Make the outermost loop iterate over cells (it should be
-    # colours). We can ignore the previous monkeypatch as this
-    # exception is encountered before the previous one.
-    monkeypatch.setattr(schedule.children[4], "_loop_type", "cells")
-
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the loop
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(schedule.children[4].loop_body[0], {"depth": 1})
-    assert ("if the parent of the supplied Loop is also a Loop then the "
-            "parent must iterate over 'colours'" in str(excinfo.value))
-
-    # Make the innermost loop iterate over cells (it should be
-    # colour). We can ignore the previous monkeypatches as this
-    # exception is encountered before the previous ones.
-    monkeypatch.setattr(schedule.children[4].loop_body[0], "_loop_type",
-                        "cells")
-
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the loop
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(schedule.children[4].loop_body[0], {"depth": 1})
-    assert ("if the parent of the supplied Loop is also a Loop then the "
-            "supplied Loop must iterate over 'cells_in_colour'"
-            in str(excinfo.value))
-
-
-def test_rc_unsupported_loop_type(monkeypatch):
-    ''' When an unsupported loop type is provided to the redundant
-    computation apply method an exception is raised. It is not
-    possible to get to this exception in normal circumstances due to
-    the validation tests so we monkey patch it. This test checks that
-    the exception is raised correctly.
-
-    '''
-    _, invoke = get_invoke("1_single_invoke.f90",
-                           TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Apply colouring
-    # Create colour transformation
-    ctrans = LFRicColourTrans()
-    # Colour the loop
-    ctrans.apply(schedule.children[4])
-
-    # Make the loop type invalid
-    monkeypatch.setattr(schedule.children[4].loop_body[0], "_loop_type",
-                        "invalid")
-
-    rc_trans = LFRicRedundantComputationTrans()
-
-    # Switch off validation
-    monkeypatch.setattr(rc_trans, "validate",
-                        lambda loop, options: None)
-
-    # Apply redundant computation to the loop
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(schedule.children[4].loop_body[0], {"depth": 1})
-    assert "Unsupported loop_type 'invalid' found" in str(excinfo.value)
-
-
-def test_rc_colour_no_loop_decrease():
-    ''' Test that we raise an exception if we try to reduce the size of a
-    loop halo depth when using the redundant computation
-    transformation. This is not allowed partly for simplicity but also
-    because, in the current implementation we might not decrease the
-    size of the relevant halo exchange as these can only be increased
-    with the current logic.
-
-    '''
-    _, invoke = get_invoke("1_single_invoke.f90",
-                           TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Create our colour transformation
-    ctrans = LFRicColourTrans()
-    # Colour the loop
-    ctrans.apply(schedule.children[4])
-
-    rc_trans = LFRicRedundantComputationTrans()
-    # First set our loop to redundantly compute to the level 2 halo
-    loop = schedule.children[4].loop_body[0]
-    rc_trans.apply(loop, {"depth": 2})
-    # Now try to reduce the redundant computation to the level 1 halo
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(loop, {"depth": 1})
-    assert ("supplied depth (1) must be greater than the existing halo depth "
-            "(2)") in str(excinfo.value)
-    # Second set our loop to redundantly compute to the maximum halo depth
-    rc_trans.apply(loop)
-    # Now try to reduce the redundant computation to a fixed value
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(loop, {"depth": 2})
-    assert ("loop is already set to the maximum halo depth so can't be "
-            "set to a fixed value") in str(excinfo.value)
-    # Now try to set the redundant computation to the same (max) value
-    # it is now
-    with pytest.raises(TransformationError) as excinfo:
-        rc_trans.apply(loop)
-    assert ("loop is already set to the maximum halo depth so this "
-            "transformation does nothing") in str(excinfo.value)
-
-
-def test_rc_colour(tmpdir):
-    ''' Test that we can redundantly compute over a colour in a coloured
-    loop. '''
-    psy, invoke = get_invoke("1_single_invoke.f90",
-                             TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Create our colour transformation
-    ctrans = LFRicColourTrans()
-    # Colour the loop
-    ctrans.apply(schedule.children[4])
-
-    # Create our redundant computation transformation
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the colour loop
-    rc_trans.apply(schedule.children[4].loop_body[0], {"depth": 2})
-
-    result = str(psy.gen)
-
-    assert (
-        "    if (f2_proxy%is_dirty(depth=2)) then\n"
-        "      call f2_proxy%halo_exchange(depth=2)\n"
-        "    end if\n"
-        "    if (m1_proxy%is_dirty(depth=2)) then\n"
-        "      call m1_proxy%halo_exchange(depth=2)\n"
-        "    end if\n"
-        "    if (m2_proxy%is_dirty(depth=2)) then\n"
-        "      call m2_proxy%halo_exchange(depth=2)\n"
-        "    end if\n" in result)
-    assert "    cmap => mesh%get_colour_map()\n" in result
-    assert "loop0_stop = ncolour" in result
-    assert ("last_halo_cell_all_colours = "
-            "mesh%get_last_halo_cell_all_colours()" in result)
-    assert (
-        "    do colour = loop0_start, loop0_stop, 1\n"
-        "      do cell = loop1_start, last_halo_cell_all_colours(colour,2), 1"
-        in result)
-
-    # We've requested redundant computation out to the level 2 halo
-    # but f1 is continuous and so the outermost halo depth (2) remains
-    # dirty. This means that all of the halo is dirty apart from level
-    # 1.
-    assert (
-        "    call f1_proxy%set_dirty()\n"
-        "    call f1_proxy%set_clean(1)" in result)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-
-def test_rc_max_colour(tmpdir):
-    ''' Test that we can redundantly compute over a colour to the maximum
-    depth in a coloured loop. '''
-    psy, invoke = get_invoke("1_single_invoke.f90",
-                             TEST_API, idx=0, dist_mem=True)
-    schedule = invoke.schedule
-
-    # Create our colour transformation
-    ctrans = LFRicColourTrans()
-    # Colour the loop
-    ctrans.apply(schedule.children[4])
-
-    # Create our redundant computation transformation
-    rc_trans = LFRicRedundantComputationTrans()
-    # Apply redundant computation to the colour loop out to the full
-    # halo depth
-    rc_trans.apply(schedule.children[4].loop_body[0])
-
-    result = str(psy.gen)
-    assert (
-        "    if (f2_proxy%is_dirty(depth=max_halo_depth_mesh)) then\n"
-        "      call f2_proxy%halo_exchange(depth=max_halo_depth_mesh)\n"
-        "    end if\n"
-        "    if (m1_proxy%is_dirty(depth=max_halo_depth_mesh)) then\n"
-        "      call m1_proxy%halo_exchange(depth=max_halo_depth_mesh)\n"
-        "    end if\n"
-        "    if (m2_proxy%is_dirty(depth=max_halo_depth_mesh)) then\n"
-        "      call m2_proxy%halo_exchange(depth=max_halo_depth_mesh)\n"
-        "    end if\n" in result)
-    assert "    cmap => mesh%get_colour_map()\n" in result
-    assert "loop0_stop = ncolour" in result
-    assert ("last_halo_cell_all_colours = "
-            "mesh%get_last_halo_cell_all_colours()" in result)
-    assert (
-        "    do colour = loop0_start, loop0_stop, 1\n"
-        "      do cell = loop1_start, last_halo_cell_all_colours(colour,"
-        "max_halo_depth_mesh), 1\n"
-        in result)
-
-    assert (
-        "    call f1_proxy%set_dirty()\n"
-        "    call f1_proxy%set_clean(max_halo_depth_mesh - 1)" in result)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-
 def test_colour_discontinuous():
     ''' Test that we raise an exception if we try to colour a loop
     containing a kernel that modifies a discontinuous field.
